o Renamed all macros in the format ERROR_CHECK_... to CHECK_... o Renamed all macros in the format ..._P to ...P. o Now not all objects are pointers to a LispObj structure, so   there are several new macros to get/check the type or value of   an object. macros in the format X...P assume that basic error   checking was already done and that the value is a real pointer,   for example:   #define XSTRINGP(object)	((object)->type == LispString_t)   #define STRINGP(object)	(POINTERP(object) && XSTRINGP(object)) o Now there is 3 types of integers (in a 32 bits computer): 	o fixnum	=> values that fit in 28 bits 	o longint	=> values that fit in 32 bits 	o bignum	=> values that don't fit in 32 bits o Type LispCharacter_t was renamed to LispSChar_t, to clarify that   currently only simple 8 bit characters are supported. o Type LispReal_t was renamed to the more proper name LispDFloat_t. o Type LispBigInteger_t was renamed to LispBignum_t and LispBigRatio_t   to LispBigratio_t. o Type LispTrue_t was removed, as now NIL, T and #<DOT> are constants,   and all of them have type LispNil_t. o All ocurrences of the macro STRPTR were replaced by the equivalent   macro ATOMID. o New math driver, now almost everything is done in mathimp.c, and only   very simple functions are implemented directly in math.c.   Note that mathimp.c was rewritten from scratch, with very few cut&paste   from the old version. Now it uses an intermediate structure, that must   be allocated in the stack, and this structure can hold any supported   number type, so the number type can change freely during the   calculations.   There is also a major cleanup in the math code, now only a few   "integer binop integer" functions are inline. o Code that uses integer math extensively should now show a significant   better performance since integers don't trigger gc.
  This patch is the integration of the new math driver and the code to
allow different sized objects. It is not yet what should be called the
final version as there is a lot of optimizations that can be done.
  One major change still missing is true vectors, as well as a cleanup
in the current buggy vector/array code.
  String handling should also be optmized, and cleaned. All strings must
end with a '\0' characters, to be safe to use with libc functions, and
even then, should check all code that handles strings internally to
properly handle strings with embeded nuls.
  Due to the large size of the changes, probably there are still bugs in
the new math driver, and there may be unsafe code that will crash when
trying to deference NIL or a fixnum to check the object type.
  I hope this is the last "huge patch".

+++ programs/xedit/lisp/helper.h
@@ -27,7 +27,7 @@
  * Author: Paulo César Pereira de Andrade
  */
 
-/* $XFree86: xc/programs/xedit/lisp/helper.h,v 1.11 2002/08/05 03:56:24 paulo Exp $ */
+/* $XFree86: xc/programs/xedit/lisp/helper.h,v 1.12 2002/11/08 08:00:56 paulo Exp $ */
 
 #ifndef Lisp_helper_h
 #define Lisp_helper_h
@@ -85,10 +85,10 @@ LispObj *LispLoadFile(LispObj*, int, int, int);
 void LispGetStringArgs(LispBuiltin*,
 			char**,	/* string1 */
 			char**,	/* string2 */
-			int*,	/* start1 */
-			int*,	/* end1 */
-			int*,	/* start2 */
-			int*);	/* end2 */
+			long*,	/* start1 */
+			long*,	/* end1 */
+			long*,	/* start2 */
+			long*);	/* end2 */
 
 /*
  string-trim character-bag string
